Defer `All_in_One_SEO_Pack_Sitemap::do_sitemaps`
`transition_post_status` is called for every modified post.
When deleting posts in bulk, the
`All_in_One_SEO_Pack_Sitemap::do_sitemaps` is called for every post.

This is not optimal.

A site with 20k posts takes 10 minutes to delete 10 posts in bulk,
because it takes 1 minute to generate the sitemaps for each of the
deleted posts.

Call the `All_in_One_SEO_Pack_Sitemap::do_sitemaps` method on the
`shutdown` action instead, and only once. This has the added benefit of
making sure that the sitemap generation process does not run out of
resources mid-save and does not interfere with callbacks that are called
on the same saving hooks.