Update to aioseop_class.php to fix post_type issue
Sometimes apply_post_title_format() is called on a page or custom post type. This currently causes the title in the <head> of the page to be printed as the format (e.g. <title>%page_title% | %blog_title%</title>) instead of the parsed template.

This fix corrects that issue so that the titles are rendered correctly.

There may be better ways to fix the issue, but the logic determining how a title is rendered is complicated and this seemed like the simplest fix